Realizzazione del form per i documenti

+++ catalog/src/main/java/it/catalog/service/dto/DocumentoDto.java
@@ -12,27 +12,27 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class DocumentoDto {
 
-	 private Long id;
-	    private String nome;
-	    private String path;
-	    private Double dimensione;
-	    private String autore;
-	    private String descrizione;
-	    private TipoDocumento categoria;
-	    private String lingua;
-	    private Integer versione;
-	    private StatiDocumento stato;
-	    private String origine;
-	    private boolean preferito; 
-	    private Integer rating; 
-	    private long visualizzazioni; 
-	    private boolean backup;
-	    private String note;
-	    private String estensione;
-	    private Instant lastView; 
-	    private Instant dataArchiviazione; 
-	    private Instant lastUpdate; 
-	    
-	 // aggiungiamo i tag direttamente nel DTO
-	    private Set <TagDto> tags;
+	private Long id;
+	private String nome;
+	private String path;
+	private Double dimensione;
+	private String autore;
+	private String descrizione;
+	private TipoDocumento categoria;
+	private String lingua;
+	private Integer versione;
+	private StatiDocumento stato;
+	private String origine;
+	private boolean preferito; 
+	private Double rating; 
+	private long visualizzazioni; 
+	private boolean backup;
+	private String note;
+	private String estensione;
+	private Instant lastView; 
+	private Instant dataArchiviazione; 
+	private Instant lastUpdate; 
+
+	// aggiungiamo i tag direttamente nel DTO
+	private Set <TagDto> tags;
 }
